requirements.txt
# 📧 Proton Mail Integration
aiosmtplib==3.0.2
aiofiles==24.1.0
email-validator==2.2.0
python-jose[cryptography]==3.3.0 

# 🤖 AI Pipeline Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# 🧠 Machine Learning & AI
torch>=2.5.0
transformers>=4.36.0
openai-whisper>=20231117
numpy>=1.24.3
scipy>=1.11.4

# 🎤 Audio Processing (STT)
librosa==0.10.1
soundfile==0.12.1
pyaudio==0.2.11
webrtcvad==2.0.10

# 🗣️ Text-to-Speech (TTS)
coqui-TTS==0.21.3
espeak-ng==1.51.1
pyttsx3==2.90

# 🌐 Networking & Communication
aiohttp==3.9.1
websockets==12.0
requests==2.31.0
httpx==0.25.2

# 📊 Monitoring & Metrics
prometheus-client==0.19.0
psutil==5.9.6
py-cpuinfo==9.0.0

# 🔧 Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
rich==13.7.0
typer==0.9.0
loguru==0.7.2 